+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+----------------
+rumqttc
+-----------
+- Change variants in `ClientError` to not expose dependence on flume/`SendError` (#420)
+- Revert erroring out when Subscription filter list is empty (#422)
+- Remove the ability to cancel `EventLoop.poll()` mid execution (#421)
+
+-----------
+
 ### R13
 ----------------
 rumqttc v0.13.0